starport destroysTo spaceport
When a hub or T&E sector capital gets redesignated, it should get a spaceport. Check if this causes problems (starport is designationOnly and would be the only such structure to destroy to something). This will allow players to bypass spaceport build time unless destroysTo respects buildTime,  but this is not a useful exploit as of right now. Fix for Ministry tickets https://ministry.kronosaur.com/record.hexm?id=59152 & http://ministry.kronosaur.com/record.hexm?id=75450.