Militia base becomes category:improvement
Addressess ministry ticket 71071.

If the militia base is an improvement, it will not disappear from a player's structures tab when zero labor is allocated to it (as industries do).

There should not be a problem with having an improvement in planetary designation industry tables; this already occurs, e.g. core.imperialAdministration.